highlight unknown
WebGPUReport wants to highlight when an avaiable limit is better
than the default (or worse in the case of compat)

To do that it needs to know the default limits.

This PR highlights limits it is unaware so hopefull the
defaults will be added. It doesn't error because we still
want the site to be useful for unknown limits.