Fix it saying "compatibility mode" in old Chrome
In browsers that have not yet implemented
`'core-features-and-limits'` the adapter will not show that
feature so it was mistakenly considered a "compatibility" adapter.

For now, for these older browser, check one of the limits on
the device. If it's a compatibilility limit then it's a compat
device. Otherwise it's a core device.